Frontend: Convert json to typescript class for API-call and create mutation component
The problem of using the HttpClient with get<T> is that it is just a
"type assertion" that fills the fields for the provided object if present
but the object is not a true instance of the class, just a prototype
with the same fields. However this leads to behavior where
all methods therefore don't work.
see https://basarat.gitbooks.io/typescript/content/docs/types/type-assertion.html

If you wish to make a smart getter or a converter for a Date (not a json
data type) then you need to have a deserializer. Solution found at:
https://stackoverflow.com/questions/22885995/how-do-i-initialize-a-typescript-object-with-a-json-object

This is a very verbose process but provides with total control.

In addition a mutation component is created